split evm/factory.rs into two files as the TODO suggests
